+++ src/main/kotlin/dev/aaronhowser/mods/aaron/misc/AaronDsls.kt
@@ -1,5 +1,6 @@
 package dev.aaronhowser.mods.aaron.misc
 
+import net.minecraft.core.Direction
 import net.neoforged.neoforge.client.model.generators.ModelBuilder
 import net.neoforged.neoforge.common.ModConfigSpec
 
@@ -15,6 +16,12 @@ object AaronDsls {
 		val elementBuilder = this.element()
 		elementBuilder.block()
 		return elementBuilder.end()
+	}
+
+	inline fun <T : ModelBuilder<T>> ModelBuilder<T>.ElementBuilder.face(direction: Direction, block: ModelBuilder<T>.ElementBuilder.FaceBuilder.() -> Unit) {
+		val faceBuilder = this.face(direction)
+		faceBuilder.block()
+		faceBuilder.end()
 	}
 
 }